continued performance improvements

+++ Grimoire/Features/LogCleanup/Commands/DeleteOldLogMessagesCommand.cs
@@ -19,7 +19,7 @@ public sealed class DeleteOldLogMessages
 {
     public sealed record Command : ICommand
     {
-        public DeleteMessageResult[] DeletedOldLogMessageIds { get; init; } = [];
+        public required IEnumerable<DeleteMessageResult> DeletedOldLogMessageIds { get; init; };
     }
 
     public sealed class Handler(GrimoireDbContext grimoireDbContext) : ICommandHandler<Command>
@@ -29,29 +29,32 @@ public sealed class DeleteOldLogMessages
         public async ValueTask<Unit> Handle(Command command, CancellationToken cancellationToken)
         {
             var successMessages = command.DeletedOldLogMessageIds
-            .Where(x => x.WasSuccessful == true)
-            .Select(x => x.MessageId)
-            .ToArray();
+                .Where(x => x.WasSuccessful)
+                .Select(x => x.MessageId)
+                .ToArray();
 
-            await this._grimoireDbContext.OldLogMessages
-                .WhereIdsAre(successMessages)
-                .ExecuteDeleteAsync(cancellationToken);
+            if(successMessages.Length != 0)
+                await this._grimoireDbContext.OldLogMessages
+                    .WhereIdsAre(successMessages)
+                    .ExecuteDeleteAsync(cancellationToken);
+
 
             var erroredMessages = command.DeletedOldLogMessageIds
-            .Where(x => x.WasSuccessful == false)
-            .Select(x => x.MessageId)
-            .ToArray();
+                .Where(x => !x.WasSuccessful)
+                .Select(x => x.MessageId)
+                .ToArray();
 
             if (erroredMessages.Length != 0)
             {
                 await this._grimoireDbContext.OldLogMessages
-                .WhereIdsAre(erroredMessages)
-                .ExecuteUpdateAsync(x => x.SetProperty(p => p.TimesTried, p => p.TimesTried + 1), cancellationToken);
+                    .WhereIdsAre(erroredMessages)
+                    .ExecuteUpdateAsync(x => x.SetProperty(p => p.TimesTried, p => p.TimesTried + 1), cancellationToken);
 
                 await this._grimoireDbContext.OldLogMessages
                     .Where(x => x.TimesTried >= 3)
                     .ExecuteDeleteAsync(cancellationToken);
             }
+
             await this._grimoireDbContext.SaveChangesAsync(cancellationToken);
             return Unit.Value;
         }
